Initial commit of issue_2061 for Win
